started refactoring to cards component. label updates via broadcast remain buggy though
