+++ lib/mindwendel_web/live/lane_live/index_component.ex
@@ -9,19 +9,6 @@ defmodule MindwendelWeb.LaneLive.IndexComponent do
   alias Mindwendel.Brainstormings
 
   @impl true
-  def handle_event("delete_idea", %{"id" => id}, socket) do
-    idea = Ideas.get_idea!(id)
-
-    %{current_user: current_user, brainstorming: brainstorming} = socket.assigns
-
-    if current_user.id in [idea.user_id | brainstorming.moderating_users |> Enum.map(& &1.id)] do
-      {:ok, _} = Ideas.delete_idea(idea)
-    end
-
-    # broadcast will take care of the removal from the list
-    {:noreply, socket}
-  end
-
   def handle_event("delete_lane", %{"id" => id}, socket) do
     lane = Lanes.get_lane!(id)
 
@@ -32,18 +19,6 @@ defmodule MindwendelWeb.LaneLive.IndexComponent do
     end
 
     # broadcast will take care of the removal from the list
-    {:noreply, socket}
-  end
-
-  def handle_event("like", %{"id" => id}, socket) do
-    Likes.add_like(id, socket.assigns.current_user.id)
-
-    {:noreply, socket}
-  end
-
-  def handle_event("unlike", %{"id" => id}, socket) do
-    Likes.delete_like(id, socket.assigns.current_user.id)
-
     {:noreply, socket}
   end
 
@@ -86,46 +61,6 @@ defmodule MindwendelWeb.LaneLive.IndexComponent do
     )
 
     {:noreply, socket}
-  end
-
-  def handle_event(
-        "add_idea_label_to_idea",
-        %{
-          "idea-id" => idea_id,
-          "idea-label-id" => idea_label_id
-        },
-        socket
-      ) do
-    idea = Ideas.get_idea!(idea_id)
-    idea_label = IdeaLabels.get_idea_label(idea_label_id)
-
-    case(IdeaLabels.add_idea_label_to_idea(idea, idea_label)) do
-      {:ok, _idea} ->
-        {:noreply, socket}
-
-      {:error, _changeset} ->
-        {:noreply, socket}
-    end
-  end
-
-  def handle_event(
-        "remove_idea_label_from_idea",
-        %{
-          "idea-id" => idea_id,
-          "idea-label-id" => idea_label_id
-        },
-        socket
-      ) do
-    idea = Ideas.get_idea!(idea_id)
-    idea_label = IdeaLabels.get_idea_label(idea_label_id)
-
-    case(IdeaLabels.remove_idea_label_from_idea(idea, idea_label)) do
-      {:ok, _idea} ->
-        {:noreply, socket}
-
-      {:error, _changeset} ->
-        {:noreply, socket}
-    end
   end
 
   def handle_event("sort_by_likes", %{"id" => id, "lane-id" => lane_id}, socket) do
