+++ priv/gettext/de/LC_MESSAGES/default.po
@@ -37,7 +37,7 @@ msgid "%{name} - New Idea"
 msgstr "%{name} - Neue Idee"
 
 #: lib/mindwendel_web/live/comment_live/show_component.html.heex:22
-#: lib/mindwendel_web/live/lane_live/index_component.html.heex:113
+#: lib/mindwendel_web/live/idea_live/card_component.html.heex:18
 #, elixir-autogen, elixir-format
 msgid "Are you sure you want to delete this idea?"
 msgstr "Möchtest du die Idee löschen?"
@@ -113,7 +113,7 @@ msgstr "Neues Brainstorming"
 msgid "New idea"
 msgstr "Neue Idee"
 
-#: lib/mindwendel_web/live/lane_live/index_component.html.heex:263
+#: lib/mindwendel_web/live/lane_live/index_component.html.heex:109
 #, elixir-autogen, elixir-format
 msgid "No ideas brainstormed"
 msgstr "Bisher keine Ideen vorhanden"
@@ -202,7 +202,7 @@ msgstr "Nutzername"
 
 #: lib/mindwendel_web/controllers/admin/brainstorming_html/export.html.heex:3
 #: lib/mindwendel_web/live/comment_live/show_component.html.heex:40
-#: lib/mindwendel_web/live/lane_live/index_component.html.heex:180
+#: lib/mindwendel_web/live/idea_live/card_component.html.heex:85
 #, elixir-autogen, elixir-format
 msgid "By"
 msgstr "Von"
@@ -421,7 +421,7 @@ msgstr "Name"
 #: lib/mindwendel_web/live/brainstorming_live/show.html.heex:26
 #: lib/mindwendel_web/live/brainstorming_live/show.html.heex:28
 #: lib/mindwendel_web/live/brainstorming_live/show.html.heex:114
-#: lib/mindwendel_web/live/lane_live/index_component.html.heex:284
+#: lib/mindwendel_web/live/lane_live/index_component.html.heex:130
 #, elixir-autogen, elixir-format
 msgid "New lane"
 msgstr "Neue Spalte"
@@ -455,7 +455,7 @@ msgstr "Spalte aktualisiert"
 
 #: lib/mindwendel_web/live/lane_live/index_component.html.heex:10
 #: lib/mindwendel_web/live/lane_live/index_component.html.heex:13
-#: lib/mindwendel_web/live/lane_live/index_component.html.heex:268
+#: lib/mindwendel_web/live/lane_live/index_component.html.heex:114
 #, elixir-autogen, elixir-format, fuzzy
 msgid "Add idea"
 msgstr "Neue Idee"
@@ -465,18 +465,18 @@ msgstr "Neue Idee"
 msgid "Default lane"
 msgstr "Eine Spalte"
 
-#: lib/mindwendel_web/live/lane_live/index_component.html.heex:279
+#: lib/mindwendel_web/live/lane_live/index_component.html.heex:125
 #, elixir-autogen, elixir-format
 msgid "No lanes available"
 msgstr "Keine Spalten vorhanden"
 
-#: lib/mindwendel_web/live/lane_live/index_component.html.heex:120
+#: lib/mindwendel_web/live/idea_live/card_component.html.heex:25
 #, elixir-autogen, elixir-format, fuzzy
 msgid "Edit idea"
 msgstr "Idee bearbeiten"
 
 #: lib/mindwendel_web/live/comment_live/show_component.html.heex:30
-#: lib/mindwendel_web/live/lane_live/index_component.html.heex:112
+#: lib/mindwendel_web/live/idea_live/card_component.html.heex:17
 #, elixir-autogen, elixir-format, fuzzy
 msgid "Delete idea"
 msgstr "Löschen"
@@ -491,9 +491,9 @@ msgstr "Idee aktualisiert"
 msgid "Additional Attachment"
 msgstr "Zusätzlicher Anhang"
 
+#: lib/mindwendel_web/live/idea_live/card_component.html.heex:75
 #: lib/mindwendel_web/live/idea_live/form_component.html.heex:43
 #: lib/mindwendel_web/live/idea_live/show_component.html.heex:27
-#: lib/mindwendel_web/live/lane_live/index_component.html.heex:170
 #, elixir-autogen, elixir-format
 msgid "No filename"
 msgstr "Kein Dateiname"
@@ -549,7 +549,7 @@ msgid "No comments available"
 msgstr "Keine Kommentare vorhanden"
 
 #: lib/mindwendel_web/live/brainstorming_live/show.html.heex:97
-#: lib/mindwendel_web/live/lane_live/index_component.html.heex:128
+#: lib/mindwendel_web/live/idea_live/card_component.html.heex:33
 #, elixir-autogen, elixir-format, fuzzy
 msgid "Show idea"
 msgstr "Zeige Idee"
