+++ priv/gettext/en/LC_MESSAGES/default.po
@@ -37,7 +37,7 @@ msgid "%{name} - New Idea"
 msgstr ""
 
 #: lib/mindwendel_web/live/comment_live/show_component.html.heex:22
-#: lib/mindwendel_web/live/lane_live/index_component.html.heex:113
+#: lib/mindwendel_web/live/idea_live/card_component.html.heex:18
 #, elixir-autogen, elixir-format
 msgid "Are you sure you want to delete this idea?"
 msgstr ""
@@ -113,7 +113,7 @@ msgstr ""
 msgid "New idea"
 msgstr ""
 
-#: lib/mindwendel_web/live/lane_live/index_component.html.heex:263
+#: lib/mindwendel_web/live/lane_live/index_component.html.heex:109
 #, elixir-autogen, elixir-format
 msgid "No ideas brainstormed"
 msgstr ""
@@ -202,7 +202,7 @@ msgstr ""
 
 #: lib/mindwendel_web/controllers/admin/brainstorming_html/export.html.heex:3
 #: lib/mindwendel_web/live/comment_live/show_component.html.heex:40
-#: lib/mindwendel_web/live/lane_live/index_component.html.heex:180
+#: lib/mindwendel_web/live/idea_live/card_component.html.heex:85
 #, elixir-autogen, elixir-format
 msgid "By"
 msgstr ""
@@ -421,7 +421,7 @@ msgstr ""
 #: lib/mindwendel_web/live/brainstorming_live/show.html.heex:26
 #: lib/mindwendel_web/live/brainstorming_live/show.html.heex:28
 #: lib/mindwendel_web/live/brainstorming_live/show.html.heex:114
-#: lib/mindwendel_web/live/lane_live/index_component.html.heex:284
+#: lib/mindwendel_web/live/lane_live/index_component.html.heex:130
 #, elixir-autogen, elixir-format
 msgid "New lane"
 msgstr ""
@@ -455,7 +455,7 @@ msgstr ""
 
 #: lib/mindwendel_web/live/lane_live/index_component.html.heex:10
 #: lib/mindwendel_web/live/lane_live/index_component.html.heex:13
-#: lib/mindwendel_web/live/lane_live/index_component.html.heex:268
+#: lib/mindwendel_web/live/lane_live/index_component.html.heex:114
 #, elixir-autogen, elixir-format, fuzzy
 msgid "Add idea"
 msgstr ""
@@ -465,18 +465,18 @@ msgstr ""
 msgid "Default lane"
 msgstr ""
 
-#: lib/mindwendel_web/live/lane_live/index_component.html.heex:279
+#: lib/mindwendel_web/live/lane_live/index_component.html.heex:125
 #, elixir-autogen, elixir-format
 msgid "No lanes available"
 msgstr ""
 
-#: lib/mindwendel_web/live/lane_live/index_component.html.heex:120
+#: lib/mindwendel_web/live/idea_live/card_component.html.heex:25
 #, elixir-autogen, elixir-format, fuzzy
 msgid "Edit idea"
 msgstr ""
 
 #: lib/mindwendel_web/live/comment_live/show_component.html.heex:30
-#: lib/mindwendel_web/live/lane_live/index_component.html.heex:112
+#: lib/mindwendel_web/live/idea_live/card_component.html.heex:17
 #, elixir-autogen, elixir-format, fuzzy
 msgid "Delete idea"
 msgstr ""
@@ -491,9 +491,9 @@ msgstr ""
 msgid "Additional Attachment"
 msgstr ""
 
+#: lib/mindwendel_web/live/idea_live/card_component.html.heex:75
 #: lib/mindwendel_web/live/idea_live/form_component.html.heex:43
 #: lib/mindwendel_web/live/idea_live/show_component.html.heex:27
-#: lib/mindwendel_web/live/lane_live/index_component.html.heex:170
 #, elixir-autogen, elixir-format
 msgid "No filename"
 msgstr ""
@@ -549,7 +549,7 @@ msgid "No comments available"
 msgstr ""
 
 #: lib/mindwendel_web/live/brainstorming_live/show.html.heex:97
-#: lib/mindwendel_web/live/lane_live/index_component.html.heex:128
+#: lib/mindwendel_web/live/idea_live/card_component.html.heex:33
 #, elixir-autogen, elixir-format, fuzzy
 msgid "Show idea"
 msgstr ""
